requirements.txt
fastapi==0.116.2
pydantic==2.11.9
python-dotenv==1.1.1
SQLAlchemy==2.0.43
uvicorn==0.35.0
